GRHB-238: + create separate style file
