+++ mobile/src/components/uam-configure-group/components/permissions-table/styles.ts
